refactor(ui): JitsiEmbed → @letscareer/live-session 패키지로 분리
packages/ui 는 AlertDialog·Popup·Toast·CategoryTabs 같은 무상태 프리미티브
모음인데, JitsiEmbed 는 네트워크 프로빙과 서버 PATCH(failover 재등록)까지
수행하는 도메인 기능이라 성격이 맞지 않았다. 최초 투입(2026-05-21) 당시엔
ZepEmbed 등 임베드 래퍼들과 같은 자리였으나, 2026-07-14 에 그 이웃들이
정리되면서 맥락이 사라진 상태였다.

web·mentor 두 앱이 함께 쓰므로 앱 도메인 폴더로는 내릴 수 없어
packages/live-session 신규 패키지로 옮긴다.

- 소스 코드는 수정하지 않음 (git rename 12건, 내용 변경 0)
- @jitsi/react-sdk 의존성을 packages/ui 에서 제거 → admin 의존성 그래프 경량화
- import 경로 13곳 갱신 (소스 6 + 테스트 mock 7)
- 미사용 LiveFeedbackMaterials.stories.tsx 삭제

부수 효과: @letscareer/ui 배럴이 ESM 전용 @jitsi/react-sdk 를 재export 하던
탓에 jest(CJS)가 파싱에 실패해, Jitsi 와 무관한 web 테스트 4스위트(마이페이지
개인정보·사이드네비 등, 25개 테스트)가 로드조차 되지 않고 있었다. 이번 분리로
해당 스위트가 정상 실행된다.

검증: web/mentor/admin 빌드 성공, Jitsi 관련 테스트 90개 통과
(live-session 45 + web 13 + mentor 32), typecheck 통과.
