+++ apps/web/package.json
@@ -30,6 +30,7 @@
     "@hookform/resolvers": "^3.3.4",
     "@letscareer/api": "workspace:*",
     "@letscareer/hooks": "workspace:*",
+    "@letscareer/live-session": "workspace:*",
     "@letscareer/mocks": "workspace:*",
     "@letscareer/store": "workspace:*",
     "@letscareer/types": "workspace:*",
