+++ apps/web/src/common/modal/JitsiEmbedModal.tsx
@@ -4,7 +4,7 @@ import {
   JitsiEmbed,
   LiveFeedbackMaterials,
   LiveSessionTimer,
-} from '@letscareer/ui/JitsiEmbed';
+} from '@letscareer/live-session/JitsiEmbed';
 
 import BaseModal from '@/common/modal/BaseModal';
