+++ packages/live-session/src/JitsiEmbed/LetsCareerLogo.tsx
