+++ packages/live-session/src/JitsiEmbed/index.tsx
