+++ packages/ui/src/index.ts
@@ -1,18 +1,3 @@
-export { JitsiEmbed } from './JitsiEmbed';
-export {
-  buildJitsiRoomUrl,
-  buildJitsiRoomName,
-  type BuildJitsiRoomUrlInput,
-} from './JitsiEmbed/buildRoomUrl';
-export {
-  resolveHealthyJitsiBaseUrl,
-  ensureLiveMeetingUrl,
-  probeJitsiExternalApi,
-  pickNextBase,
-  safeHost,
-  type EnsureLiveMeetingUrlOptions,
-  type EnsureLiveMeetingUrlResult,
-} from './JitsiEmbed/jitsiHealthCheck';
 export * from './AlertDialog';
 export * from './Popup';
 export * from './Toast';
